Add new file for Justin (#3)
* Add new file for Justin

Above is the subject line, this is extended details.
Look at these clues:
Don’t end your commit message with a period.
Keep your commit messages to 50 characters or less. Add extra detail in the extended description window if necessary. This is located just below the subject line.
Use active voice. For example, "add" instead of "added" and "merge" instead of "merged".
Think of your commit as expressing intent to introduce a change.

* Update 0000-01-02-JustinFenton.md

Add a meme